Extract Run report formatter helpers into taskpane-run-report-formatters
Move five pure formatting helpers out of taskpane.js into a new dedicated
module to reduce taskpane.js size and keep Run report logic easier to maintain.

Moved: runReportSkipReasonLabel, runReportStatusLabel, runReportMetricTypes,
formatIssueDetailsForReport, runReportWarningDetails.

BATCH_SKIP_REASONS import removed from taskpane.js (now only used in the
new module). No behavior or report shape changes.
